requirements.txt
asgiref==3.5.2
aws-psycopg2==1.3.8
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==4.0.6
django-allauth==0.51.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
future==0.16.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.8.5
pycparser==2.21
PyJWT==2.4.0
# pypiwin32==223
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
# pywin32==304
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
simplejson==3.17.6
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.15.1
wcwidth==0.1.9
gunicorn==20.1.0